+++ src/Resources/app/administration/src/module/lengow-connector/component/lgw-free-trial-warning/index.js
@@ -7,24 +7,27 @@ const { Component } = Shopware;
 Component.register('lgw-free-trial-warning', {
     template,
 
-    inject: ['LengowConnectorSyncService'],
+    props: {
+        accountStatusData: {
+            type: Object,
+            required: true
+        }
+    },
 
     data() {
         return {
-            freeTrialEnabled: false,
             dayLeft: '',
-            isExpired: false,
             link: LENGOW_URL
         };
     },
 
     created() {
-        this.LengowConnectorSyncService.getAccountStatus(false).then(result => {
-            if (result.success) {
-                this.freeTrialEnabled = result.type === 'free_trial';
-                this.dayLeft = result.day;
-                this.isExpired = result.expired;
-            }
-        });
+        this.createdComponent();
+    },
+
+    methods: {
+        createdComponent() {
+            this.dayLeft = this.accountStatusData.day;
+        }
     }
 });