+++ src/Resources/app/administration/src/module/lengow-connector/page/lgw-contact/index.js
@@ -4,5 +4,29 @@ import './lgw-contact.scss';
 const { Component } = Shopware;
 
 Component.register('lgw-contact', {
-    template
+    template,
+
+    inject: ['LengowConnectorSyncService'],
+
+    data() {
+        return {
+            helpCenterLink: '',
+            supportLink: ''
+        };
+    },
+
+    created() {
+        this.createdComponent();
+    },
+
+    methods: {
+        createdComponent() {
+            this.LengowConnectorSyncService.getPluginLinks().then(result => {
+                if (result.success) {
+                    this.helpCenterLink = result.links.help_center;
+                    this.supportLink = result.links.support;
+                }
+            });
+        }
+    }
 });